Rename fpx to fpx-cli and fpx-lib to fpx
