+++ fpx-cli/src/api.rs
@@ -1,0 +1,2 @@
+pub mod client;
+pub mod errors;
