+++ fpx-cli/src/commands.rs
