+++ fpx-cli/src/commands/client.rs
