+++ fpx-cli/src/commands/client/spans.rs
