+++ fpx-cli/src/commands/client/traces.rs
