+++ fpx-cli/src/commands/debug.rs
