+++ fpx-cli/src/commands/dev.rs
@@ -3,7 +3,7 @@ use crate::events::InMemoryEvents;
 use crate::grpc::GrpcService;
 use crate::initialize_fpx_dir;
 use anyhow::{Context, Result};
-use fpx_lib::{api, service};
+use fpx::{api, service};
 use opentelemetry_proto::tonic::collector::trace::v1::trace_service_server::TraceServiceServer;
 use std::future::IntoFuture;
 use std::path::PathBuf;
