+++ fpx-cli/src/commands/system.rs
