+++ fpx-cli/src/commands/system/database.rs
