+++ fpx-cli/src/data/util.rs
@@ -1,0 +1,38 @@
+use fpx::data::{DbError, Result};
+use libsql::{de, Rows};
+use serde::de::DeserializeOwned;
+
+#[allow(dead_code)]
+pub(crate) trait RowsExt {
+    /// `T` must be a `struct`
+    async fn fetch_one<T: DeserializeOwned>(&mut self) -> Result<T, DbError>;
+
+    /// `T` must be a `struct`
+    async fn fetch_optional<T: DeserializeOwned>(&mut self) -> Result<Option<T>, DbError>;
+
+    /// `T` must be a `struct`
+    async fn fetch_all<T: DeserializeOwned>(&mut self) -> Result<Vec<T>, DbError>;
+}
+
+impl RowsExt for Rows {
+    async fn fetch_one<T: DeserializeOwned>(&mut self) -> Result<T, DbError> {
+        self.fetch_optional().await?.ok_or(DbError::NotFound)
+    }
+
+    async fn fetch_optional<T: DeserializeOwned>(&mut self) -> Result<Option<T>, DbError> {
+        match self.next().await? {
+            Some(row) => Ok(Some(de::from_row(&row)?)),
+            None => Ok(None),
+        }
+    }
+
+    async fn fetch_all<T: DeserializeOwned>(&mut self) -> Result<Vec<T>, DbError> {
+        let mut results = Vec::new();
+
+        while let Some(row) = self.next().await? {
+            results.push(de::from_row(&row)?);
+        }
+
+        Ok(results)
+    }
+}
