--- fpx-lib/Cargo.toml
@@ -1,61 +1,0 @@
-[package]
-name = "fpx-lib"
-version = { workspace = true }
-edition = "2021"
-authors = { workspace = true }
-license = { workspace = true }
-repository = { workspace = true }
-
-[lib]
-
-[dependencies]
-anyhow = { version = "1.0", default-features = false }
-async-trait = { version = "0.1", default-features = false }
-axum = { workspace = true, default-features = false, features = [
-    "json",
-    "matched-path",
-] }
-bytes = { version = "1.0", default-features = false }
-fpx-macros = { version = "0.1.0", path = "../fpx-macros" }
-futures-util = { version = "0.3", default-features = false }
-hex = { version = "0.4", default-features = false, features = ["alloc"] }
-http = { version = "1.1", default-features = false }
-libsql = { version = "0.6", default-features = false, optional = true }
-opentelemetry = { version = "0.27", default-features = false }
-opentelemetry_sdk = { version = "0.27", default-features = false }
-opentelemetry-proto = { version = "0.27", default-features = false, features = [
-    "trace",
-    "gen-tonic-messages",
-    "with-serde",
-    "with-schemars",
-] }
-prost = { version = "0.13", default-features = false }
-schemars = { workspace = true, default-features = false, features = ["derive"] }
-serde = { version = "1.0", default-features = false, features = ["derive"] }
-serde_json = { version = "1.0", default-features = false }
-strum = { version = "0.26", default-features = false, features = ["derive"] }
-thiserror = { version = "2.0", default-features = false }
-time = { version = "0.3", default-features = false, features = [
-    "serde-human-readable",
-] }
-tokio = { version = "1.40", default-features = false }
-tower = { version = "0.4", default-features = false }
-tower-http = { version = "0.5", default-features = false, features = [
-    "compression-br",
-    "compression-gzip",
-    "decompression-br",
-    "decompression-gzip",
-] }
-tracing = { version = "0.1", default-features = false, features = [
-    "attributes",
-] }
-tracing-opentelemetry = { version = "0.28", default-features = false }
-url = { version = "2.5", default-features = false }
-wasm-bindgen = { version = "0.2", default-features = false, optional = true }
-
-[dev-dependencies]
-http-body-util = { version = "0.1", default-features = false }
-tokio = { version = "1.40", default-features = false, features = [
-    "macros",
-    "test-util",
-] }
