--- fpx-lib/README.md
@@ -1,6 +1,0 @@
-# fpx-lib
-
-This crate contains the shared types and logic that is used in both the normal
-fpx binary and the fpx-workers crates. This crate contains most of the logic
-that is shared between them and it also includes shared traits. The
-implementations of these traits might not be included in this crate.
