--- fpx-lib/src/events.rs
@@ -1,7 +1,0 @@
-use crate::api::models::ServerMessage;
-use axum::async_trait;
-
-#[async_trait]
-pub trait ServerEvents: Sync + Send {
-    async fn broadcast(&self, msg: ServerMessage);
-}
