+++ fpx-workers/Cargo.toml
@@ -19,7 +19,7 @@ crate-type = ["cdylib"]
 [dependencies]
 axum = { workspace = true, default-features = false }
 console_error_panic_hook = { version = "0.1" }
-fpx-lib = { version = "0.1.0", path = "../fpx-lib", features = [
+fpx = { version = "0.1.0", path = "../fpx", features = [
     "wasm-bindgen",
 ] }
 getrandom = { version = "0.2", features = ["js"] } # Required for its feature
