+++ fpx/Cargo.toml
@@ -6,78 +6,56 @@ authors = { workspace = true }
 license = { workspace = true }
 repository = { workspace = true }
 
-[features]
-embed-studio = [] # When enabled it will embed Studio from frontend/dist
+[lib]
 
 [dependencies]
-anyhow = { workspace = true }
-async-trait = { version = "0.1" }
+anyhow = { version = "1.0", default-features = false }
+async-trait = { version = "0.1", default-features = false }
 axum = { workspace = true, default-features = false, features = [
-    "http1",
-    "query",
-    "tokio",
-    "tracing",
-    "ws",
+    "json",
+    "matched-path",
 ] }
-bytes = { version = "1.6" }
-clap = { workspace = true, features = ["derive", "env"] }
-fpx-lib = { version = "0.1.0", path = "../fpx-lib", features = ["libsql"] }
+bytes = { version = "1.0", default-features = false }
 fpx-macros = { version = "0.1.0", path = "../fpx-macros" }
-futures-util = { version = "0.3" }
-hex = { version = "0.4" }
-http = { version = "1.1" }
-http-body-util = { version = "0.1" }
-include_dir = { version = "0.7.3" }
-libsql = { version = "0.6", default-features = false, features = [
-    "core",
-    "serde",
-] }
-once_cell = { version = "1.19" }
-opentelemetry = { version = "0.27" }
-opentelemetry_sdk = { version = "0.27", features = ["rt-tokio"] }
-opentelemetry-otlp = { version = "0.27", features = [
-    "http-json",
-    "reqwest-client",
-    "reqwest-rustls-webpki-roots",
-] }
-opentelemetry-proto = { version = "0.27", features = [
+futures-util = { version = "0.3", default-features = false }
+hex = { version = "0.4", default-features = false, features = ["alloc"] }
+http = { version = "1.1", default-features = false }
+libsql = { version = "0.6", default-features = false, optional = true }
+opentelemetry = { version = "0.27", default-features = false }
+opentelemetry_sdk = { version = "0.27", default-features = false }
+opentelemetry-proto = { version = "0.27", default-features = false, features = [
+    "trace",
     "gen-tonic-messages",
     "with-serde",
     "with-schemars",
 ] }
-prost = { version = "0.13" }
-rand = { version = "0.8.5" }
-reqwest = { version = "0.12", default-features = false, features = [
-    "charset",
-    "http2",
-    "rustls-tls",
-    "json",
+prost = { version = "0.13", default-features = false }
+schemars = { workspace = true, default-features = false, features = ["derive"] }
+serde = { version = "1.0", default-features = false, features = ["derive"] }
+serde_json = { version = "1.0", default-features = false }
+strum = { version = "0.26", default-features = false, features = ["derive"] }
+thiserror = { version = "2.0", default-features = false }
+time = { version = "0.3", default-features = false, features = [
+    "serde-human-readable",
 ] }
-schemars = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-serde_with = { version = "3.8.1" }
-strum = { version = "0.26", features = ["derive"] }
-thiserror = { version = "2.0" }
-time = { version = "0.3.17", features = ["serde-human-readable"] }
-tokio = { version = "1.40", features = ["rt-multi-thread", "signal", "fs"] }
-tokio-tungstenite = { version = "0.21", features = [
-    "rustls-tls-webpki-roots",
-] } # This should be kept the same as whatever Axum has
-toml = { version = "0.8" }
-tonic = { version = "0.12" }
-tower = { version = "0.4" }
-tracing = { version = "0.1" }
-tracing-opentelemetry = { version = "0.28" }
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-url = { version = "2.5" }
-
-[target.'cfg(windows)'.dependencies]
-libsql = { version = "0.6", default-features = false, features = [
-    "core",
-    "serde",
-    "replication",
+tokio = { version = "1.40", default-features = false }
+tower = { version = "0.4", default-features = false }
+tower-http = { version = "0.5", default-features = false, features = [
+    "compression-br",
+    "compression-gzip",
+    "decompression-br",
+    "decompression-gzip",
 ] }
+tracing = { version = "0.1", default-features = false, features = [
+    "attributes",
+] }
+tracing-opentelemetry = { version = "0.28", default-features = false }
+url = { version = "2.5", default-features = false }
+wasm-bindgen = { version = "0.2", default-features = false, optional = true }
 
 [dev-dependencies]
-test-log = { version = "0.2", default-features = false, features = ["trace"] }
+http-body-util = { version = "0.1", default-features = false }
+tokio = { version = "1.40", default-features = false, features = [
+    "macros",
+    "test-util",
+] }
