+++ xtask/Cargo.toml
@@ -9,7 +9,7 @@ repository = { workspace = true }
 [dependencies]
 anyhow = { workspace = true }
 clap = { workspace = true, features = ["derive", "env"] }
-fpx-lib = { version = "0.1.0", path = "../fpx-lib" }
+fpx = { version = "0.1.0", path = "../fpx" }
 schemars = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
